General improvements to the editor (#7)
Now supports block indent and unindent.
Now supports block commenting and uncommenting.
Now when you press return for a new line, it will start indented at the previous level.
Now when you press backspace to remove the indentation it removes the indentation in blocks of multiples of 4 spaces.
Now when you press new line after a : character it will auto indent the next line by and extra 4 characters.
Tests have been written for these changes, and they reproduced by use simulating the key. presses. They don't cover every detail that was implemented but cover the majority.

Saving a script now defaults to a .py extension, (no tests are written for this.)

Indenting and unindenting alters the text to the nearest indent in multiples of four spaces.

When indenting or unindenting lines with tabs, it should convert them to spaces.

I've also tweaked the editor colours, making the selected line highlighting less bright, and increasing the comment brightness a bit.